requirements.txt
msal>=1.28.0
requests>=2.32.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.2.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
